# OpenAPI_v1.yaml
openapi: 3.0.0
info:
  title: Media Distribution System API
  version: 1.0.0
  description: >
    API для управления устройствами (Orange Pi), загрузки медиафайлов,
    синхронизации контента и работы с личным кабинетом администратора.
servers:
  - url: https://server.com/api
paths:
...
  /admin/devices/{deviceId}:
    get:
      summary: Данные конкретного устройства
      tags: [Admin Devices]
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"

    patch:
      summary: Изменить параметры устройства
      tags: [Admin Devices]
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        "200":
          description: OK
...
  /admin/files/{fileId}:
    get:
      summary: Получить метаданные файла
      tags: [Admin Files]
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/File"

    delete:
      summary: Удалить файл
      tags: [Admin Files]
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted


  ############################
  # DEVICE AUTH
  ############################
...
components:
  schemas:
    Device:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        status: { type: string }
        last_sync: { type: string, format: date-time }
    File:
      type: object
      properties:
        id: { type: integer }
        filename: { type: string }
        size: { type: integer }
        hash: { type: string }
...